refactor(protocol): separate wire contract domains
