+++ config/architecture-budgets.json
@@ -8,7 +8,6 @@
     "crates/connect-agent/src/server.rs": 1967,
     "crates/connect-cli/src/lib.rs": 1815,
     "crates/connect-mirror/src/lib.rs": 1915,
-    "crates/connect-protocol/src/lib.rs": 1589,
     "packages/sync/src/adoption.ts": 1127,
     "packages/sync/src/index.ts": 1116,
     "packages/sync/src/mirror.ts": 1255,
